+++ nextgen/project-graph/src/projects_locator.rs
@@ -20,41 +20,57 @@ pub fn infer_project_id_and_source(path: &str) -> miette::Result<(Id, WorkspaceR
 
 /// For each pattern in the globs list, glob the file system
 /// for potential projects, and infer their name and source.
-pub fn locate_projects_with_globs<I, V>(
+pub fn locate_projects_with_globs<'glob, I, V>(
     workspace_root: &Path,
     globs: I,
     sources: &mut FxHashMap<Id, WorkspaceRelativePathBuf>,
     vcs: Option<&BoxedVcs>,
 ) -> miette::Result<()>
 where
-    I: IntoIterator<Item = V>,
-    V: AsRef<str>,
+    I: IntoIterator<Item = &'glob V>,
+    V: AsRef<str> + 'glob,
 {
-    let root_source = ".".to_owned();
-    let globs = globs
-        .into_iter()
-        .map(|glob| glob.as_ref().to_owned())
-        .collect::<Vec<_>>();
+    let mut locate_globs = vec![];
 
     // Root-level project has special handling
-    if globs.contains(&root_source) {
-        let root_id = fs::file_name(workspace_root);
+    for glob in globs.into_iter() {
+        let glob = glob.as_ref();
 
-        sources.insert(
-            Id::clean(if root_id.is_empty() {
-                "root"
-            } else {
-                root_id.as_str()
-            })?,
-            WorkspaceRelativePathBuf::from(root_source),
-        );
+        if glob == "." {
+            let root_id = fs::file_name(workspace_root);
+
+            sources.insert(
+                Id::clean(if root_id.is_empty() {
+                    "root"
+                } else {
+                    root_id.as_str()
+                })?,
+                WorkspaceRelativePathBuf::from("."),
+            );
+        } else {
+            locate_globs.push(glob);
+        }
     }
 
     // Glob for all other projects
-    let mut potential_projects = glob::walk(workspace_root, &globs)?;
+    let mut potential_projects = glob::walk(workspace_root, locate_globs)?;
     potential_projects.sort();
 
-    for project_root in potential_projects {
+    for mut project_root in potential_projects {
+        // Remove trailing moon.yml
+        if project_root.is_file() {
+            if project_root.ends_with(consts::CONFIG_PROJECT_FILENAME) {
+                project_root = project_root.parent().unwrap().to_owned();
+            } else {
+                warn!(
+                    source = ?project_root,
+                    "Received a file path for a project root, must be a directory",
+                );
+
+                continue;
+            }
+        }
+
         if project_root.is_dir() {
             let project_source =
                 to_virtual_string(project_root.strip_prefix(workspace_root).unwrap())?;